+++ NEWS.md
@@ -3,6 +3,7 @@
 ### Enhancements
 
 * Updated the package docs and vignettes with the new way of specifying data for `teal::init()`. The `data` argument will accept a `teal_data` object
+* Outlier labels no longer appear out of bounds in `tm_a_regression`.
 
 ### Miscellaneous